==> notebooks/00-ll-dataTreatment.ipynb <==
import pandas as pd

cooperados = pd.read_csv("C:/Users/leall/Downloads/ListaCooperados.csv", sep=';' , encoding='utf-8')
cooperados

cooperados['CpfCnpj'] = cooperados['CpfCnpj'].astype(str)

cooperados.info()

errados = cooperados[cooperados['CpfCnpj'].str.len() <= 10]

quantidade_registros = len(errados)


print("Linhas filtradas:")
print(errados)

contagem_tipo_pessoa = errados['TipoPessoa'].value_counts()

contagem_tipo_pessoa

cpf = cooperados[cooperados['CpfCnpj'].str.len() == 11]

quantidade_cpf = len(cpf)

print(quantidade_cpf)

x = cpf['TipoPessoa'].value_counts()

cpf.to_csv("cooperados_corrigido.csv")

cnpj = cooperados[cooperados['CpfCnpj'].str.len() == 14]

quantidade_cnpj = len(cnpj)

print(quantidade_cnpj)

y = cnpj['TipoPessoa'].value_counts()

y

cooperados['cnpj'] = cooperados.loc[cooperados['TipoPessoa'] == 'J', 'CpfCnpj'].fillna('-')

cooperados['cpf'] = cooperados.loc[cooperados['TipoPessoa'] == 'F', 'CpfCnpj'].fillna('-')

cooperados['cpf'] = cooperados['cpf'].fillna('-')
cooperados['cnpj'] = cooperados['cnpj'].fillna('-')

cooperados = cooperados.drop(columns=['CpfCnpj'])

cooperados.to_csv("corrigido_cnpj.csv")

fix_data = pd.read_csv("C:/Users/leall/Downloads/operacao_credito_estadual_att.csv", sep=';')

fix_data

fix_data['DT_EMISSAO'] = pd.to_datetime(fix_data['DT_EMISSAO'],format='mixed').dt.strftime('%d/%m/%Y')
fix_data['DT_VENCIMENTO'] = pd.to_datetime(fix_data['DT_VENCIMENTO'],format='mixed').dt.strftime('%d/%m/%Y')
fix_data['DT_FIM_COLHEITA'] = pd.to_datetime(fix_data['DT_FIM_COLHEITA'],format='mixed').dt.strftime('%d/%m/%Y')
fix_data['DT_INIC_COLHEITA'] = pd.to_datetime(fix_data['DT_INIC_COLHEITA'],format='mixed').dt.strftime('%d/%m/%Y')
fix_data['DT_INIC_PLANTIO'] = pd.to_datetime(fix_data['DT_INIC_PLANTIO'],format='mixed').dt.strftime('%d/%m/%Y')
fix_data['DT_COMUNICACAO'] = pd.to_datetime(fix_data['DT_COMUNICACAO'],format='mixed').dt.strftime('%d/%m/%Y')
fix_data['DT_FIM_PLANTIO'] = pd.to_datetime(fix_data['DT_FIM_PLANTIO'],format='mixed').dt.strftime('%d/%m/%Y')

print(fix_data)

fix_data.shape

fix_data = fix_data.drop_duplicates()

fix_data.shape

fix_data.to_csv("op_estadual_att.csv")
