==> notebooks/00-ll-firstExploration.ipynb <==
import pandas as pd

produtos = pd.read_csv(r'../../data/Produto.csv', encoding='ISO-8859-1')

soja = produtos[produtos['DESCRICAO'] == 'SOJA']
soja

empreendimento = pd.read_csv(r'../../data/Empreendimento.csv', sep=';', encoding='ISO-8859-1')
cod_empreendimento =  empreendimento[empreendimento['PRODUTO'] == 'SOJA']
cod_empreendimento

operacao_credito_estadual = pd.read_csv(r'C:/Users/leall/OneDrive/Área de Trabalho/02_TABS_BASICAS_OPERACAO_CREDITO_RURAL_PROAGRO_RECURSOS_PUBLICOS_PRIVADOS/SICOR_OPERACAO_BASICA_ESTADO_2018_ATUAL.csv', sep=';', encoding='ISO-8859-1')
operacao_credito_estadual_antes = pd.read_csv(r'C:/Users/leall/OneDrive/Área de Trabalho/02_TABS_BASICAS_OPERACAO_CREDITO_RURAL_PROAGRO_RECURSOS_PUBLICOS_PRIVADOS/SICOR_OPERACAO_BASICA_ESTADO_2013_2017.csv', sep=';', encoding='ISO-8859-1')


operacao_credito_estadual_total =pd.concat([operacao_credito_estadual,operacao_credito_estadual_antes], axis=0)

ref_bacen = operacao_credito_estadual_total[operacao_credito_estadual_total['CD_EMPREENDIMENTO'].isin(set(cod_empreendimento['CODIGO']))]


sicor_cop_basico = pd.read_csv("C:/Users/leall/OneDrive/Área de Trabalho/02_TABS_BASICAS_OPERACAO_CREDITO_RURAL_PROAGRO_RECURSOS_PUBLICOS_PRIVADOS/SICOR_COP_BASICO.csv", sep=';', encoding='ISO-8859-1')
sicor_cop_basico = sicor_cop_basico.drop(columns=['DT_FIM_COLHEITA', 'DT_FIM_PLANTIO', 'DT_INICIO_PLANTIO', 'DT_INICIO_COLHEITA'])
sicor_cop_basico


merged = ref_bacen.merge(sicor_cop_basico, on=['REF_BACEN', 'NU_ORDEM'], how='inner')

merged

glebas = pd.read_csv(r'C:/Users/leall/OneDrive/Área de Trabalho/03_TABS_COMP_BASICAS_OPERACOES_CREDITO_RURAL_PROAGRO_RECURSOS_PUB/Glebas.csv', sep=';', encoding='ISO-8859-1')
glebas_soja = glebas[(glebas['REF_BACEN'].isin(set(merged['REF_BACEN']))) & (glebas['NU_ORDEM'].isin(set(merged['NU_ORDEM'])))]
glebas_soja


count_registers = glebas_soja['REF_BACEN'].value_counts()

codigos_a_remover = count_registers[count_registers <= 3].index.tolist()

final = glebas_soja[~glebas_soja['REF_BACEN'].isin(codigos_a_remover)]



sicor_op_glebas = pd.merge(final, merged, on=['REF_BACEN', 'NU_ORDEM'], how='left')
sicor_op_glebas

relatorio_glebas = sicor_op_glebas[['REF_BACEN',	'NU_ORDEM',	'DT_EMISSAO',	'DT_VENCIMENTO',	'CD_ESTADO',	'CD_EMPREENDIMENTO',	'CD_PROGRAMA'	,'CD_TIPO_IRRIGACAO'	,'CD_TIPO_AGRICULTURA',	'CD_FASE_CICLO_PRODUCAO',	'CD_TIPO_CULTIVO',	'CD_TIPO_GRAO_SEMENTE',	'DT_FIM_COLHEITA',	'DT_FIM_PLANTIO','DT_INIC_COLHEITA','DT_INIC_PLANTIO','DT_COMUNICACAO','CD_CICLO_CULTIVAR','CD_TIPO_SOLO',	'CD_EVENTO']]
relatorio_glebas
