==> notebooks/00-ll-inserts.ipynb <==
import pandas as pd

df = pd.read_csv("C:/Users/leall/OneDrive/Área de Trabalho/api-lara/GeoForesight-back/notebooks/op_estadual_att.csv", sep=';', encoding='utf-8')
df = df.drop_duplicates()
df.shape

with open('C:/Users/leall/OneDrive/Área de Trabalho/api-lara/GeoForesight-back/database/insert_postgres/insert_op.sql', 'w') as file:

    for index, row in df.iterrows():
        ref_bacen = str(row['ref_bacen'])
        nu_ordem = str(row['nu_ordem'])
        inicio_plantio = row['inicio_plantio']
        final_plantio = row['final_plantio']
        inicio_colheita = row['inicio_colheita']
        final_colheita = row['final_colheita']
        data_vencimento = row['data_vencimento']
        idempreendimento = row['idempreendimento']
        idevento = row['idevento']
        idsolo = row['idsolo']
        idirrigacao = row['idirrigacao']
        idciclo = row['idciclo']
        idgrao = row['idgrao']
        idcultivar = row['idcultivar']
        idprograma = row['idprograma']
        estado = row['estado']

        
        dataframe = query = f"INSERT INTO PUBLIC.operacao_credito_estadual(ref_bacen, nu_ordem, inicio_plantio, final_plantio, inicio_colheita, final_colheita, data_vencimento, idempreendimento, idevento, idsolo, idirrigacao, idciclo, idgrao, idcultivar, idprograma, estado) VALUES ('{ref_bacen}', '{nu_ordem}', '{inicio_plantio}', '{final_plantio}', '{inicio_colheita}', '{final_colheita}', '{data_vencimento}', {idempreendimento}, {idevento}, {idsolo}, {idirrigacao}, {idciclo}, {idgrao}, {idcultivar}, {idprograma}, '{estado}');\n"
        file.write(query)




glebas_df = pd.read_csv("C:/Users/leall/OneDrive/Área de Trabalho/glebas_ref.csv", sep=';', encoding='ISO-8859-1')
glebas_df


codigo_para_excluir = [
'5891994',
'9150493',
'511648794',
'513613305',
'513311173',
'509447347',
'510124977',
'511735340',
'509309030',
'508023326',
'513613305',
'506007208',
'506742889',
'9339233',
'513524122',
'512291209',
'510760955',
'513847914',
'511500680',
'509831558',
'506777883',
'505876386',
'512405342',
'510278401',
'512261191',
'506925664',
'512461666',
'512405342',
'513735136',
'512943606',
'512363012',
'508967472',
'507240703',
'512701247',
'512342231',
'513166147',
'510102339',
'508222205',
'1642834',
'512531354',
'NULL82',
'5891994']

arquivo_entrada = "C:/Users/leall/OneDrive/Área de Trabalho/api-lara/GeoForesight-back/data/INSERT ATT_TESTE.sql"
arquivo_saida = "C:/Users/leall/OneDrive/Área de Trabalho/api-lara/GeoForesight-back/database/insert_postgres/inserts_glebas.sql"

comandos_sql_sem_exclusao = []
with open(arquivo_entrada, 'r') as arquivo_entrada:
    linhas = arquivo_entrada.readlines()
    for linha in linhas:
        if not any(codigo in linha for codigo in codigo_para_excluir):
            comandos_sql_sem_exclusao.append(linha)
with open(arquivo_saida, 'w') as arquivo_saida:
    arquivo_saida.writelines(comandos_sql_sem_exclusao)

print(f"Registros com códigos {codigo_para_excluir} excluídos com sucesso.")

municipios = pd.read_csv("C:/Users/leall/OneDrive/Documentos/municipio_glebas.csv", sep=';', encoding='ISO-8859-1')
municipios.dropna()
municipios.drop_duplicates()
municipios.shape[0]


municipios_glebas = municipios.merge(glebas_df,on=['REF_BACEN','NU_ORDEM'])

with open('../database/insert_postgres/insert_municipio.sql', 'w', encoding='utf-8') as file:
    for index, row in municipios_glebas.iterrows():
        ref_bacen = str(row['REF_BACEN'])
        nu_ordem = str(row['NU_ORDEM'])
        nome = str(row['nome'])

        query1 = f"INSERT INTO PUBLIC.municipio_glebas(ref_bacen, nu_ordem, nome) VALUES ('{ref_bacen}','{nu_ordem}','{nome}');\n"
        file.write(query1)
