==> notebooks/01-lg-criptoTest.ipynb <==
!pip install bcrypt

import bcrypt

# Senha a ser criptografada
senha = "admin123"

# Gerar um salt aleatório
salt = bcrypt.gensalt()

# Criptografar a senha com o salt
hashed_password = bcrypt.hashpw(senha.encode('utf-8'), salt)

print(senha)
# Imprimir o valor de hash
print(hashed_password.decode('utf-8'))

hashed_password = hashed_password.decode('utf-8')


###########################
### comparacao de senha ###
###########################

# Suponha que você tenha recuperado o valor da senha criptografada do banco de dados
senha_armazenada_no_banco = b'$2b$12$.9DID1fRTcqshr0kJWtaQOLt8K4r2Y5.LCZSnmL0gBiwq5rC7aGN.'

# Suponha que você tenha recebido a senha fornecida pelo usuário durante o login
senha_fornecida_pelo_usuario = "admin123"

# Compare a senha fornecida com a senha armazenada no banco de dados
if bcrypt.checkpw(senha_fornecida_pelo_usuario.encode('utf-8'), senha_armazenada_no_banco):
    # Senhas coincidem, login bem-sucedido
    print("Login bem-sucedido!")
else:
    # Senhas não coincidem, autenticação falha
    print("Credenciais inválidas.")
