==> notebooks/2.1-lg-criptoMongo.ipynb <==
pip install -r ../requirements.txt

from pymongo import MongoClient
import bcrypt

uri = "mongodb+srv://phantom:<password>@cluster0.yxkoek8.mongodb.net/?retryWrites=true&w=majority"
client = MongoClient(uri)

# Nome da coleção
collection = client.GeoForesight.User

# Suponha que você tenha o email e senha fornecidos no seu código
email = "admin@admin.com"
senha_fornecida = "admin123"

# Encontre o usuário com o email fornecido
user = collection.find_one({"email": email})

# Recupere o hash da senha do documento MongoDB
hashed_password = user.get("senha")

# Compare a senha fornecida com a senha armazenada no banco de dados
if bcrypt.checkpw(senha_fornecida.encode('utf-8'), hashed_password.encode('utf-8')):
    # Senhas coincidem, login bem-sucedido
    print("Login bem-sucedido!")
else:
    # Senhas não coincidem, autenticação falha
    print("Credenciais inválidas.")


# Feche a conexão com o MongoDB
client.close()
