==> notebooks/3.0-lg-outputEmail.ipynb <==
import pandas as pd

dict_teste = [
  {
    "aceitacao_email": True,
    "data_aceitacao": "Sun, 29 Oct 2023 16:25:20 GMT",
    "id_user": 2,
    "u.email": "lguirald@admin.com"
  },
  {
    "aceitacao_email": True,
    "data_aceitacao": "Mon, 30 Oct 2023 14:16:04 GMT",
    "id_user": 3,
    "u.email": "marcus@email.com"
  },
  {
    "aceitacao_email": True,
    "data_aceitacao": "Mon, 30 Oct 2023 17:38:20 GMT",
    "id_user": 4,
    "u.email": "lara.o.leal2001@gmail.com"
  }
]

lista_de_emails = []
for item in dict_teste:
    if 'u.email' in item:
        lista_de_emails.append(item['u.email'])

lista_de_emails
