==> notebooks/4.0-lg-Forecast-New.ipynb <==
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import json
import os

pd.set_option('display.max_rows', 500)

from statsmodels.tsa.statespace.sarimax import SARIMAX

dados = pd.DataFrame(columns=['ref_bacen', 'date', 'indice'])

diretorio = '../data/serieT/'

arquivos_json = [arquivo for arquivo in os.listdir(diretorio) if arquivo.endswith('.json')]

for arquivo_json in arquivos_json:
    with open(os.path.join(diretorio, arquivo_json), 'r') as file:
        d = json.loads(file.read())
        
        data_temp = pd.DataFrame({
            'ref_bacen': arquivo_json,
            'date': d['Indices']['NDVI']['Serie Processada']['Data'],
            'indice': d['Indices']['NDVI']['Serie Processada']['Indice']
        })

        dados = pd.concat([dados, data_temp], ignore_index=True)

dados['ref_bacen'] = dados['ref_bacen'].str.split('.').str.get(0)

dados['date'] = pd.to_datetime(dados['date'], format='%Y-%m-%d')
dados = dados.set_index('date')

dados.dtypes

dados[dados['ref_bacen'] == '511332955']

Model validation    

previsoes = []

for ref_bacen in dados['ref_bacen'].unique():
    print(f"Processando ref_bacen: {ref_bacen}")
    dados_ref = dados[dados['ref_bacen'] == ref_bacen][['indice']]  # Seleciona como DataFrame com uma única coluna

    modelo = SARIMAX(dados_ref, order=(1, 1, 1), seasonal_order=(1, 1, 1, 12))
    
    try:
        resultado = modelo.fit()

        # Obter os próximos 6 passos de previsão a partir do próximo mês após a última data nos dados
        ultima_data = dados_ref.index[-1]
        proximas_datas_previsao = pd.date_range(start=ultima_data, periods=7, freq='M')[1:]

        previsao = resultado.get_forecast(steps=6)
        previsao_mean = previsao.predicted_mean
        
        previsao_df = pd.DataFrame({
            'date': proximas_datas_previsao,
            'ref_bacen': ref_bacen,
            'previsao': previsao_mean,
        })

        previsoes.append(previsao_df)

    except Exception as e:
        print(f"Erro para ref_bacen {ref_bacen}: {e}")

previsoes_df = pd.concat(previsoes)


previsoes_df['date'] = pd.to_datetime(previsoes_df['date'], format='%Y-%m-%d')
final_df = pd.merge(dados, previsoes_df, how='outer', on=['date', 'ref_bacen'])
final_df['date'] = final_df['date'].dt.strftime('%Y-%m')

final_df = final_df.drop_duplicates(subset=['date', 'ref_bacen'], keep='first')

final_df = final_df[final_df['ref_bacen']!='st']

final_df[final_df['ref_bacen'] == '512309065']

final_df.to_csv('../data/previsoes_1.csv')
