==> notebooks/4.1-lg-Forecast.ipynb <==
from sqlalchemy import create_engine, text
from flask import Flask, jsonify, request
from dotenv import load_dotenv
import pandas as pd
import os

load_dotenv()

df_previsoes = pd.read_csv('../data/previsoes.csv')
df_previsoes = df_previsoes.loc[:, ~df_previsoes.columns.str.startswith('Unnamed')]

df_prev = pd.read_csv('../data/previsoes_1.csv')
df_prev = df_prev.loc[:, ~df_prev.columns.str.startswith('Unnamed')]

concatenado = pd.concat([df_previsoes, df_prev])
concatenado

concatenado.head()

concatenado.to_csv('../data/conc.csv')

engine = create_engine(os.getenv('url_heroku'))

# nome_da_tabela = 'table_prev_orig'

# concatenado.to_sql(nome_da_tabela, engine, index=False, if_exists='replace')

# print(f"Tabela '{nome_da_tabela}' criada com sucesso no PostgreSQL.")

criar consulta dos dados na view 'public.vw_query_glebas_resultado'

query = ''' SELECT * from 
        vw_query_glebas_resultado
            WHERE 1=1'''

engine = create_engine(os.getenv('url_heroku'))

conn = engine.connect()

# Executar a query
resultados = conn.execute(text(query)).fetchall()

# Montar a lista de resultados
lista_resultados = []
for resultado in resultados:
    resultado_dict = {
        "ref_bacen": resultado.ref_bacen,
        "nu_identificador": resultado.nu_identificador,
        "coordenadas": resultado.coordenadas,
        "inicio_plantio": resultado.inicio_plantio,
        "final_plantio": resultado.final_plantio,
        "inicio_colheita": resultado.inicio_colheita,
        "final_colheita": resultado.final_colheita,
        "descricao_grao": resultado.descricao_grao,
        "descricao_irrigacao": resultado.descricao_irrigacao,
        "descricao_producao": resultado.descricao_producao,
        "descricao_solo": resultado.descricao_solo, 
        "descricao_evento": resultado.descricao_evento, 
        "descricao_cultiva": resultado.descricao_cultiva,
        "estado": resultado.estado,
        "municipio": resultado.municipio,
        "produto": resultado.produto
    }
    lista_resultados.append(resultado_dict)

#  Fechar a conexão
conn.close()

df = pd.DataFrame(lista_resultados)

df

predicoes_df = df.merge(concatenado, on='ref_bacen', how='left')

predicoes_df['previsao']

predicoes_df[~predicoes_df['date'].isna()].shape

aplicar modelo de previsão de indices

engine = create_engine(os.getenv('url_heroku'))
conn = engine.connect()

predicoes_df.to_sql('predicoes_v2', conn, if_exists='replace', index=False)

salvar a nova coluna de previsões na views
